requirements.txt
absl-py                 
albumentations         
anyio               
anylabeling           
argon2-cffi             
argon2-cffi-bindings   
attrs                  
backcall                
beautifulsoup4       
bleach               
cachetools             
certifi              
cffi                 
charset-normalizer      
colorama              
coloredlogs           
comm                  
cycler                
darkdetect             
debugpy                
decorator             
defusedxml              
entrypoints        
et-xmlfile            
exceptiongroup        
faiss-cpu            
fastjsonschema          
filelock              
flatbuffers           
fonttools            
gdown                
gitdb                  
GitPython             
google-auth           
google-auth-oauthlib    
grpcio                
h5py                   
humanfriendly           
idna                
imageio           
imgaug                
imgviz                  
importlib-metadata   
importlib-resources   
ipykernel               
ipython                
ipython-genutils        
ipywidgets             
jedi                  
Jinja2                 
joblib             
jsonschema           
jupyter                
jupyter_client         
jupyter-console       
jupyter_core           
jupyter-server         
jupyterlab-pygments    
jupyterlab-widgets     
kiwisolver            
labelme                
lxml                   
Markdown               
MarkupSafe              
matplotlib            
matplotlib-inline      
mistune             
mpmath                 
natsort               
nbclassic            
nbclient              
nbconvert            
nbformat               
nest-asyncio            
networkx              
notebook               
notebook_shim           
numpy                   
oauthlib            
onnx                
onnxruntime           
opencv-contrib-python  
opencv-python          
opencv-python-headless
openpyxl            
packaging             
pandas                
pandocfilters           
parso                
pickleshare            
Pillow        
pip                    
pkgutil_resolve_name   
prometheus-client     
prompt-toolkit    
protobuf              
psutil               
py-cpuinfo       
pyasn1                 
pyasn1-modules         
PyAudio               
pycocotools           
pycparser              
Pygments             
pyparsing             
pyproject-toml        
PyQt5                   
PyQt5-Qt5             
PyQt5-sip             
pyreadline        
pyrsistent          
PySocks              
python-dateutil        
pytz                    
PyWavelets              
pywin32                
pywinpty               
PyYAML                 
pyzmq               
qimage2ndarray         
qtconsole  
QtPy                    
qudida                  
requests                
requests-oauthlib       
rsa                     
scikit-image            
scikit-learn           
scipy                   
seaborn                 
Send2Trash              
setuptools              
shapely                 
six                     
smmap                  
sniffio                
soundfile               
soupsieve               
supervision             
sympy                   
tensorboard           
tensorboard-data-server 
tensorboard-plugin-wit  
termcolor               
terminado              
threadpoolctl           
tifffile               
tinycss2               
toml                   
torch              
torchaudio          
torchvision           
tornado                 
tqdm                    
traitlets            
typing_extensions      
ultralytics            
urllib3                 
wcwidth                 
webencodings            
websocket-client        
Werkzeug              
wheel                   
widgetsnbextension   
wincertstore            
zipp                   